pyproject.toml
[project]
name = "quanta-forecast"
version = "0.1.0"
description = "Multi-model time series forecasting toolkit — classical + ML + deep + foundation"
authors = [{ name = "Rushil Kaul", email = "theruka007@outlook.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"

dependencies = [
    "numpy>=2.1",
    "polars>=1.12",
    "pandas>=2.2",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "rich>=13.9",
    "plotly>=5.24",
    "typer>=0.13",
]

[project.optional-dependencies]
# `classical` powers ARIMA / SARIMA via statsmodels. statsforecast +
# prophet are included as alternative classical backends users can register
# themselves; the built-in `arima` backend depends only on statsmodels.
classical = [
    "statsmodels>=0.14",
    "scipy>=1.14",
    "statsforecast>=1.7",
    "prophet>=1.1",
]
# `ml` powers the LightGBM backends. mlforecast is kept for users who want
# the Nixtla ML stack; the built-in `lightgbm` backend only needs lightgbm.
ml = [
    "lightgbm>=4.5",
    "mlforecast>=0.14",
]
deep = [
    "darts>=0.30",
    "neuralforecast>=1.7",
]
foundation = [
    "transformers>=4.46",
    "chronos-forecasting>=1.4",
    "torch>=2.5",
]
demo = [
    "streamlit>=1.40",
    "yfinance>=0.2",
]
dev = [
    "ruff>=0.8",
    "mypy>=1.13",
    "pytest>=8.3",
    "pytest-cov>=6.0",
    # httpx is pulled in by fastapi.testclient; making it explicit so
    # `uv sync --extra dev` on a fresh checkout runs the API tests.
    "httpx>=0.27",
]

[project.scripts]
quanta = "quanta.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/quanta"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
# Intentional relaxations:
# * E501  — docstrings and comments occasionally run long; black-style
#           line wrap is enforced by `ruff format` separately.
# * B008  — typer idiomatically calls `typer.Option(...)` in defaults.
# * B904  — HTTPException chaining is noise in small API handlers.
# * N803  — SessionLocal-style capitalized names are a SQLAlchemy idiom.
ignore = ["E501", "B008", "B904", "N803", "N806"]

[tool.ruff.lint.per-file-ignores]
# Tests can shadow builtins and use private names of the modules they test.
"tests/*.py" = ["N802", "N803", "N806"]

[tool.mypy]
python_version = "3.13"
# Not strict by default — we pass a lot of numpy + statsmodels output
# through, which defeats strict typing without adding value. Tighten on a
# per-module basis when helpful.
ignore_missing_imports = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
# --no-cov on CI; keep coverage when developers run pytest locally.
addopts = "-ra --cov=quanta --cov-report=term-missing"
